[refac] #37 dto 구조 수정
