+++ src/main/java/org/hilingual/domain/voca/api/service/VocaService.java
@@ -8,7 +8,6 @@ import org.hilingual.domain.voca.core.facade.VocaRetriever;
 import org.hilingual.domain.voca.core.facade.VocaSearcher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.hilingual.domain.voca.api.dto.res.VocaSearchResponse;
 
 
 import java.util.List;
@@ -30,10 +29,6 @@ public class VocaService {
     // 단어장 검색
     public VocaSearchListResponse searchVocaList(final Long userId, final String keyword) {
         final List<Voca> vocas = vocaSearcher.searchStartsWith(userId, keyword);
-        return new VocaSearchListResponse(
-                vocas.stream()
-                        .map(VocaSearchResponse::from)
-                        .toList()
-        );
+        return VocaSearchListResponse.from(vocas);
     }
 }
